# openapi.json
{"openapi": "3.1.0", "info": {"title": "VKT Bot API", "version": "1.0.0"}, "paths": {"/api/auth/login": {"post": {"tags": ["auth"], "summary": "Login", "description": "Login and get JWT token.", "operationId": "login_api_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/auth/me": {"get": {"tags": ["auth"], "summary": "Get Current User Info", "description": "Get current authenticated user information.", "operationId": "get_current_user_info_api_auth_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/users": {"get": {"tags": ["users"], "summary": "List Users", "description": "List all users with pagination. Admin only.", "operationId": "list_users_api_users_get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "default": 1, "title": "Page"}}, {"name": "size", "in": "query", "required": false, "schema": {"type": "integer", "default": 20, "title": "Size"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedUsersResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["users"], "summary": "Create User", "description": "Create new user. Admin only.", "operationId": "create_user_api_users_post", "security": [{"HTTPBearer": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/users/{username}": {"get": {"tags": ["users"], "summary": "Get User", "description": "Get user by username. Admin only.", "operationId": "get_user_api_users__username__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string", "title": "Username"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "patch": {"tags": ["users"], "summary": "Update User", "description": "Update user. Admin only.", "operationId": "update_user_api_users__username__patch", "security": [{"HTTPBearer": []}], "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string", "title": "Username"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["users"], "summary": "Delete User", "description": "Delete user. Admin only.", "operationId": "delete_user_api_users__username__delete", "security": [{"HTTPBearer": []}], "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string", "title": "Username"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/chats": {"get": {"tags": ["chats"], "summary": "List Chats", "description": "List all chats with pagination. Admin only.", "operationId": "list_chats_api_chats_get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "default": 1, "title": "Page"}}, {"name": "size", "in": "query", "required": false, "schema": {"type": "integer", "default": 20, "title": "Size"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedChatsResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/chats/{chat_id}": {"get": {"tags": ["chats"], "summary": "Get Chat", "description": "Get chat by ID. Admin only.", "operationId": "get_chat_api_chats__chat_id__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "chat_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Chat Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/chats/{chat_id}/send-message": {"post": {"tags": ["chats"], "summary": "Send Message", "description": "Send a message to a chat. Admin only.", "operationId": "send_message_api_chats__chat_id__send_message_post", "security": [{"HTTPBearer": []}], "parameters": [{"name": "chat_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Chat Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SendMessageRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SendMessageResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/roles": {"get": {"tags": ["roles"], "summary": "List Roles", "description": "List all roles with pagination. Admin only.", "operationId": "list_roles_api_roles_get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "default": 1, "title": "Page"}}, {"name": "size", "in": "query", "required": false, "schema": {"type": "integer", "default": 20, "title": "Size"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedRolesResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["roles"], "summary": "Create Role", "description": "Create new role. Admin only.", "operationId": "create_role_api_roles_post", "security": [{"HTTPBearer": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RoleCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RoleResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/roles/{role_id}": {"patch": {"tags": ["roles"], "summary": "Update Role", "description": "Update role. Admin only.", "operationId": "update_role_api_roles__role_id__patch", "security": [{"HTTPBearer": []}], "parameters": [{"name": "role_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Role Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RoleUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RoleResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["roles"], "summary": "Delete Role", "description": "Delete role. Admin only.", "operationId": "delete_role_api_roles__role_id__delete", "security": [{"HTTPBearer": []}], "parameters": [{"name": "role_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Role Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/chat-users": {"get": {"tags": ["chat-users"], "summary": "List Chat Users", "description": "List all chat users with pagination. Admin only.", "operationId": "list_chat_users_api_chat_users_get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "default": 1, "title": "Page"}}, {"name": "size", "in": "query", "required": false, "schema": {"type": "integer", "default": 20, "title": "Size"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedChatUsersResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/chat-users/{user_id}": {"get": {"tags": ["chat-users"], "summary": "Get Chat User", "description": "Get chat user by ID with roles and chats. Admin only.", "operationId": "get_chat_user_api_chat_users__user_id__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatUserDetailResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/chat-users/{user_id}/roles/{role_id}": {"post": {"tags": ["chat-users"], "summary": "Assign Role To User", "description": "Assign role to chat user. Admin only.", "operationId": "assign_role_to_user_api_chat_users__user_id__roles__role_id__post", "security": [{"HTTPBearer": []}], "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}, {"name": "role_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Role Id"}}], "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "object", "additionalProperties": {"type": "string"}, "title": "Response Assign Role To User Api Chat Users  User Id  Roles  Role Id  Post"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["chat-users"], "summary": "Remove Role From User", "description": "Remove role from chat user. Admin only.", "operationId": "remove_role_from_user_api_chat_users__user_id__roles__role_id__delete", "security": [{"HTTPBearer": []}], "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}, {"name": "role_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Role Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/bot-settings/{key}": {"get": {"tags": ["bot-settings"], "summary": "Get Bot Setting", "description": "Get bot setting by key. Admin only.", "operationId": "get_bot_setting_api_bot_settings__key__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "key", "in": "path", "required": true, "schema": {"type": "string", "title": "Key"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BotSettingsResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["bot-settings"], "summary": "Update Bot Setting", "description": "Update or create bot setting. Admin only.", "operationId": "update_bot_setting_api_bot_settings__key__put", "security": [{"HTTPBearer": []}], "parameters": [{"name": "key", "in": "path", "required": true, "schema": {"type": "string", "title": "Key"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateBotSettingsRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BotSettingsResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/bot-settings": {"get": {"tags": ["bot-settings"], "summary": "List Bot Settings", "description": "List all bot settings. Admin only.", "operationId": "list_bot_settings_api_bot_settings_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/BotSettingsResponse"}, "type": "array", "title": "Response List Bot Settings Api Bot Settings Get"}}}}}, "security": [{"HTTPBearer": []}]}}, "/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/health": {"get": {"summary": "Health", "operationId": "health_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/gl/webhooks": {"get": {"tags": ["gitlab"], "summary": "List Webhooks", "description": "List all GitLab webhooks with pagination.", "operationId": "list_webhooks_gl_webhooks_get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "default": 1, "title": "Page"}}, {"name": "size", "in": "query", "required": false, "schema": {"type": "integer", "default": 10, "title": "Size"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GlWebhookListResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["gitlab"], "summary": "Create Webhook", "description": "Create a new GitLab webhook.", "operationId": "create_webhook_gl_webhooks_post", "security": [{"HTTPBearer": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GlWebhookCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GlWebhookRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/gl/webhooks/{webhook_id}": {"get": {"tags": ["gitlab"], "summary": "Get Webhook", "description": "Get a specific GitLab webhook by ID.", "operationId": "get_webhook_gl_webhooks__webhook_id__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "webhook_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Webhook Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GlWebhookRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "patch": {"tags": ["gitlab"], "summary": "Update Webhook", "description": "Update a GitLab webhook.", "operationId": "update_webhook_gl_webhooks__webhook_id__patch", "security": [{"HTTPBearer": []}], "parameters": [{"name": "webhook_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Webhook Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GlWebhookUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GlWebhookRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["gitlab"], "summary": "Delete Webhook", "description": "Delete a GitLab webhook.", "operationId": "delete_webhook_gl_webhooks__webhook_id__delete", "security": [{"HTTPBearer": []}], "parameters": [{"name": "webhook_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Webhook Id"}}], "responses": {"204": {"description": "Successful Response"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/gl/webhooks/{webhook_id}/trigger": {"post": {"tags": ["gitlab"], "summary": "Trigger Webhook", "description": "GitLab webhook trigger endpoint.", "operationId": "trigger_webhook_gl_webhooks__webhook_id__trigger_post", "parameters": [{"name": "webhook_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Webhook Id"}}, {"name": "x-gitlab-token", "in": "header", "required": true, "schema": {"type": "string", "title": "X-Gitlab-Token"}}, {"name": "x-gitlab-event", "in": "header", "required": true, "schema": {"type": "string", "title": "X-Gitlab-Event"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": true, "title": "Data"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"BotSettingsResponse": {"properties": {"key": {"type": "string", "title": "Key"}, "value": {"type": "string", "title": "Value"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["key", "value", "updated_at"], "title": "BotSettingsResponse", "description": "Bot settings response schema."}, "ChatResponse": {"properties": {"id": {"type": "string", "title": "Id"}, "title": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Title"}}, "type": "object", "required": ["id"], "title": "ChatResponse"}, "ChatUserChatResponse": {"properties": {"id": {"type": "string", "title": "Id"}, "type": {"type": "string", "title": "Type"}}, "type": "object", "required": ["id", "type"], "title": "ChatUserChatResponse"}, "ChatUserDetailResponse": {"properties": {"id": {"type": "string", "title": "Id"}, "roles": {"items": {"$ref": "#/components/schemas/ChatUserRoleResponse"}, "type": "array", "title": "Roles"}, "chats": {"items": {"$ref": "#/components/schemas/ChatUserChatResponse"}, "type": "array", "title": "Chats"}}, "type": "object", "required": ["id", "roles", "chats"], "title": "ChatUserDetailResponse"}, "ChatUserResponse": {"properties": {"id": {"type": "string", "title": "Id"}}, "type": "object", "required": ["id"], "title": "ChatUserResponse"}, "ChatUserRoleResponse": {"properties": {"id": {"type": "string", "title": "Id"}, "name": {"type": "string", "title": "Name"}}, "type": "object", "required": ["id", "name"], "title": "ChatUserRoleResponse"}, "GlWebhookCreate": {"properties": {"name": {"type": "string", "title": "Name", "default": ""}, "secret": {"type": "string", "title": "Secret"}, "chat_id": {"type": "string", "title": "Chat Id"}}, "type": "object", "required": ["secret", "chat_id"], "title": "GlWebhookCreate", "description": "Create GitLab Webhook schema."}, "GlWebhookListResponse": {"properties": {"items": {"items": {"$ref": "#/components/schemas/GlWebhookRead"}, "type": "array", "title": "Items"}, "total": {"type": "integer", "title": "Total"}, "page": {"type": "integer", "title": "Page"}, "size": {"type": "integer", "title": "Size"}, "pages": {"type": "integer", "title": "Pages"}}, "type": "object", "required": ["items", "total", "page", "size", "pages"], "title": "GlWebhookListResponse", "description": "List response for webhooks."}, "GlWebhookRead": {"properties": {"name": {"type": "string", "title": "Name", "default": ""}, "id": {"type": "string", "format": "uuid", "title": "Id"}, "chat_id": {"type": "string", "title": "Chat Id"}, "chat_title": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Chat Title"}, "created_by_id": {"type": "string", "title": "Created By Id"}, "created_by_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Created By Name"}, "last_used_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Last Used At"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["id", "chat_id", "created_by_id", "last_used_at", "created_at", "updated_at"], "title": "GlWebhookRead", "description": "Read GitLab Webhook schema."}, "GlWebhookUpdate": {"properties": {"name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}}, "type": "object", "title": "GlWebhookUpdate", "description": "Update GitLab Webhook schema."}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "LoginRequest": {"properties": {"username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["username", "password"], "title": "LoginRequest"}, "PaginatedChatUsersResponse": {"properties": {"items": {"items": {"$ref": "#/components/schemas/ChatUserResponse"}, "type": "array", "title": "Items"}, "total": {"type": "integer", "title": "Total"}, "page": {"type": "integer", "title": "Page"}, "size": {"type": "integer", "title": "Size"}, "pages": {"type": "integer", "title": "Pages"}}, "type": "object", "required": ["items", "total", "page", "size", "pages"], "title": "PaginatedChatUsersResponse"}, "PaginatedChatsResponse": {"properties": {"items": {"items": {"$ref": "#/components/schemas/ChatResponse"}, "type": "array", "title": "Items"}, "total": {"type": "integer", "title": "Total"}, "page": {"type": "integer", "title": "Page"}, "size": {"type": "integer", "title": "Size"}, "pages": {"type": "integer", "title": "Pages"}}, "type": "object", "required": ["items", "total", "page", "size", "pages"], "title": "PaginatedChatsResponse"}, "PaginatedRolesResponse": {"properties": {"items": {"items": {"$ref": "#/components/schemas/RoleResponse"}, "type": "array", "title": "Items"}, "total": {"type": "integer", "title": "Total"}, "page": {"type": "integer", "title": "Page"}, "size": {"type": "integer", "title": "Size"}, "pages": {"type": "integer", "title": "Pages"}}, "type": "object", "required": ["items", "total", "page", "size", "pages"], "title": "PaginatedRolesResponse"}, "PaginatedUsersResponse": {"properties": {"items": {"items": {"$ref": "#/components/schemas/UserResponse"}, "type": "array", "title": "Items"}, "total": {"type": "integer", "title": "Total"}, "page": {"type": "integer", "title": "Page"}, "size": {"type": "integer", "title": "Size"}, "pages": {"type": "integer", "title": "Pages"}}, "type": "object", "required": ["items", "total", "page", "size", "pages"], "title": "PaginatedUsersResponse"}, "RoleCreate": {"properties": {"name": {"type": "string", "title": "Name"}}, "type": "object", "required": ["name"], "title": "RoleCreate"}, "RoleResponse": {"properties": {"name": {"type": "string", "title": "Name"}, "id": {"type": "string", "format": "uuid", "title": "Id"}}, "type": "object", "required": ["name", "id"], "title": "RoleResponse"}, "RoleUpdate": {"properties": {"name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}}, "type": "object", "title": "RoleUpdate"}, "SendMessageRequest": {"properties": {"text": {"type": "string", "maxLength": 4096, "minLength": 1, "title": "Text", "description": "Message text to send"}, "parse_mode": {"anyOf": [{"type": "string", "enum": ["MarkdownV2", "HTML"]}, {"type": "null"}], "title": "Parse Mode", "description": "Parse mode for message formatting"}}, "type": "object", "required": ["text"], "title": "SendMessageRequest"}, "SendMessageResponse": {"properties": {"success": {"type": "boolean", "title": "Success"}, "message": {"type": "string", "title": "Message"}}, "type": "object", "required": ["success", "message"], "title": "SendMessageResponse"}, "Token": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type", "default": "bearer"}}, "type": "object", "required": ["access_token"], "title": "Token"}, "UpdateBotSettingsRequest": {"properties": {"value": {"type": "string", "title": "Value"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}}, "type": "object", "required": ["value"], "title": "UpdateBotSettingsRequest", "description": "Update bot settings request schema."}, "UserCreate": {"properties": {"username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}, "is_superuser": {"type": "boolean", "title": "Is Superuser", "default": false}, "is_active": {"type": "boolean", "title": "Is Active", "default": true}}, "type": "object", "required": ["username", "password"], "title": "UserCreate"}, "UserResponse": {"properties": {"username": {"type": "string", "title": "Username"}, "email": {"type": "string", "title": "Email"}, "is_superuser": {"type": "boolean", "title": "Is Superuser"}, "is_active": {"type": "boolean", "title": "Is Active"}}, "type": "object", "required": ["username", "email", "is_superuser", "is_active"], "title": "UserResponse"}, "UserUpdate": {"properties": {"username": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Username"}, "password": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Password"}, "is_superuser": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Is Superuser"}, "is_active": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Is Active"}}, "type": "object", "title": "UserUpdate"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"HTTPBearer": {"type": "http", "scheme": "bearer"}}}}